+++ conf.py
@@ -140,12 +140,6 @@ CodeChat_lexer_for_glob = {
     '.gitignore': 'Perl',
 }
 
-# **CodeChat note:** _`CodeChat_excludes` is a list of exclude_patterns_ which
-# applies only to source documents; in contrast, exclude_patterns_ will exclude
-# the given files from all of Sphinx (for example, exclude_patterns_ files
-# won't be included even if they're mentioned in html_static_path_).
-CodeChat_excludes = []
-
 # `source_encoding <http://sphinx-doc.org/config.html#confval-source_encoding>`_:
 # The encoding of source files.
 ##source_encoding = 'utf-8-sig'